refactor : add favorites
